When serializing objects with expand, convert QuerySets to lists.
Serializing objects and expanding fields stored the fields as QuerySets
(when appropriate), but this meant that if something modified that list
before it was returned to the client, the data could change. It wasn't a
suitable snapshot. Now we convert to a list during serialization,
preserving the value at that moment in time.