Add Extension-level configurability of metadata.
Extensions were taking the parent package's metadata, but this meant
that a package could only contain one extension, which is annoying.
Technically, they could contain more, but each would get the name,
version, etc. of the main package itself.

Now, an Extension can define a metadata dictionary that contains the
information. The keys/values are in the form of standard PyPI package
metadata. This will let us stay consistent in how we represent metadata
across all packages, which will be important later for an extension
store.