Improve several of the widgets in jquery.gravy.
autoSizeTextArea no longer leaves the proxy elements sitting around in
in the <body> tag when it's destroyed.

inlineEditor can be told not to focus a textarea by default by setting
focusOnOpen to false.

modalBox can be told to place itself in a container other than <body>.

modalBox can also take an ID for the outer modalbox container.

funcQueue now takes an optional context for the callbacks.