Allow extensions to specify their list of app directories.
Extensions can define an 'apps' list as part of their class definition.
When loading, all items listed in apps will be added to INSTALLED_APPS,
and when unloading, those will all be removed.

If an extension does not provide 'apps', then this falls back on the
default behavior of using the app containing the extension class.