Add a URL link for extension authors, and make the page more stylable.
The extensions list page now links the author to the extension's URL, if
one is available.

The name, version and author are now grouped together in a div that can
be styled. The page also has a class so that decisions can be made
involving the overall style of the extensions list page.

There's also now an author_url field (mapping to 'Author-home-page')
that extensions can provide in Extension.metadata for specifying the
URL for the author's site, for the link. By default, this will point to
the extension's URL (which is generally the project page for the
extension).