Add utilities for the javascript message catalog.
Django has a view for serving a javascript implementation of gettext and a
message catalog for the desired language. This view recomputes everything every
time it's fetched, which is kind of terrible. This change adds some utilities
for caching the data in a way which works like our old media serial did.

Testing done:
Used this in conjunction with some mocked-up javascript i18n files for
different languages and saw it generate and use good cache keys.

Reviewed at https://reviews.reviewboard.org/r/4451/