Use extension package name rather than display name for the media dir.
Extensions have two different "Name" fields--one of which comes from the
packaging, and is often short and ugly, and one of which is defined in the
extension metadata dictionary, and is meant for display to the user. Since the
introduction of the metadata overrides, we were using the latter for the name
of the media dir, when we wanted to be using the former.

Testing done:
Enabled an extension in reviewboard with an overridden "Name" field and saw
that the media directory in htdocs/media/ext was created using the package
name.

Reviewed at https://reviews.reviewboard.org/r/4461/