For edit icons on inline editors, use CSS visibility instead of display.
We try to use CSS min-height on the labels for inline editors to keep them the
same height when the edit icon disappears, which works for the most part,
except many browsers think that the text baseline moves, which still makes the
text jump by a pixel or two.

This change alters our javascript to manipulate the visibility parameter of the
icon rather than just calling .show() and .hide(), which retains space for the
element on the page.

Testing done:
Opened and closed a bunch of inline editors and saw it be silky smooth.

Reviewed at https://reviews.reviewboard.org/r/4503/