Change django.conf.urls.defaults -> django.conf.urls
Django 1.4 changed the namespace for 'patterns' and 'url' from
django.conf.urls.defaults to django.conf.urls. This was listed in the
deprecation notes, but we never updated it because 1.4 maintained
compatibility. 1.5 got rid of the compatibility shim, so we need to actually
update now.

Testing done:
* Ran unit tests
* Used in conjunction with review board

Reviewed at https://reviews.reviewboard.org/r/4570/