Adjust the way that we override HttpResponse.content.
The WebAPIResponse.content property overrides Django's HttpResponse.content in
kind of a hacky way, using private APIs that have gone away in Django 1.5. This
change overrides the property by proxying back in to the content property in
the super class. This is kind of awkward because of the way that properties
work, but the method herein seems to work and is much less gross than the
previous approach.

Testing done: Ran unit tests
Reviewed at https://reviews.reviewboard.org/r/4583/