Default to no indentation for the json templatetag.
The json_dumps templatetag has an "indent" parameter, which defaulted to '0'.
This caused the generated output to include newlines, but indent things zero
spaces. If we default this to None, it will skip all extraneous whitespace.

Testing done:
Used this in a future change, and saw that it now dumped my big data blob onto
a single line in the output rather than formatting it super ugly.

Reviewed at https://reviews.reviewboard.org/r/4605/