Increased z-index on modal-box.
A prior change to Review Board made it so the user account navigation pop-up
was showing over a modal-box. Increase the z-index to put everything on the
right layer.

Testing done:
Clicked on a review request, hit review which brought up the modal screen and
ensured that the usernav drop down was no longer visible over the modal box.

Reviewed at https://reviews.reviewboard.org/r/4607/