Convert datagrids to use a spritesheet for icons and add retina support.
This change moves the artwork for the datagrids into a spritesheet which is
rendered at both 1x and 2x resolution. This works in a very similar way to the
Review Board artwork.

Testing done:
Checked the appearance of all the new icons on both low-dpi and high-dpi
screens.

Reviewed at https://reviews.reviewboard.org/r/4672/