Split up extensions/base.py into several files.
Back when the extension support was smaller, everything went into
base.py. This become harder to work with as things grew, so now it's
split up into multiple files.

The orignal base.py still exists, and imports the main classes that were
used previously. It's there for backwards-compatibility, but will be
removed in a future version.