Remove backwards-compatibility for allowed_{item,list}_mimetypes.
In 0.6.x, we deprecated the allowed_{item,list}_mimetypes variables, in
support of the consolidated allowed_mimetypes list. For compatibility,
we built up the new list from the old ones, if they were set. While not
a ton of work, it was extra work we didn't need to do, and extra
storage.

Support for those old lists are now gone. Clients must use
allowed_mimetypes.

Along with this, allowed_mimetypes entries no longer need to have both
an 'item' and a 'list' entry. They default to None.