Bump AJAX_SERIAL when bumping the Extension sync generation.
Extensions often affect many aspects of page rendering. Pages that make
use of AJAX_SERIAL in their cache keys expect that a change to a
template will cause a change to the cache key. It therefore makes sense
to bump AJAX_SERIAL when we bump the extension sync generation (which
happens when enabling/disabling an extension or changing extension
configuration).

With this change, any pages that properly use AJAX_SERIAL will
invalidate their cache any time an extension makes a change that may,
for example, toggle a feature on or off.