Allow inlineEditor dirty field calculation to be overridden.
The current dirty field calculation in inlineEditor is fine when dealing
with textareas, which have no concept of dirty state to begin with, but
custom text fields may be smarter than that. In those cases, we want
callers to be able to compute a dirty state on their own.

Now, options.isFieldDirty can be provided to return whether or not an
editor is dirty. The default implementation compares text, as it's
always done.