Use DjangoJSONEncoder in the json_dumps template tag.
By using DjangoJSONEncoder, we can properly serialize a few extra types
that the default JSONEncoder does not handle the way that's typically
expected in a Django app, such as timestamps. This ensures that a
timestamp serialized using json_dumps will appear the same as a
timestamp serialized through a WebAPIResource.