Fix URL errors when configuring extensions in the presence of SITE_ROOT
The extension framework used str.lstrip() to try to strip off the SITE_ROOT
from the admin URL root, but the problem with that is that lstrip takes in a
set of characters, not a string. If that set of characters included "a" and "d"
(for example, if SITE_ROOT is "/reviewboard/"), it would remove those
characters from the beginning of "admin/...".

Reviewed at https://reviews.reviewboard.org/r/4892/