Improve and fix a couple things in inlineEditor.
This adds a 'title' attribute to the "required" asterisk that says "This
field is required."

It also fixes a bug with setting the width of the editor when the
content wraps. We were previously setting min-width, and I believe this
was from before we properly fixed all the width calculations (to ensure
we had a minimum, and didn't have it shrink unintentionally). Removing
the min-width setting doesn't seem to negatively impact any cases where
we've used an inlineEditor.