Remove a bunch of now-unused stuff from djblets.auth.
Some of this hasn't been used in a long time, and some of it is only unused now
that I made changes to the Review Board code to use django.contrib.auth's
versions. I'm hoping to get rid of this package entirely, but registration
still has to be dealt with.

Testing done:
* Ran djblets unit tests.
* Ran Review Board unit tests.
* Smoke tested authentication pages (log in, log out, user prefs, etc).

Reviewed at https://reviews.reviewboard.org/r/4938/