Split up files in djblets.util into new apps.
djblets.util has become a dumping grounds of all sorts of random stuff.
In an effort to make things more manageable and future-proof, this
change moves out a lot of our code into new apps.

This change is backwards-compatible. The old files can still be imported
and will import the old classes and functions they provide, but will do
so with deprecation warnings. In Djblets 0.9, these will be removed.

There's also a few PyFlakes fixes I fixed while in the area.