Expand the amount of information in ExtensionResource.
ExtensionResource now lists additional information on the extension,
such as the summary, version, and related links. This will be used for
a slightly improved Manage Extensions page, and could be used for
command line management tools.

There's also an optimization thrown in for fetching the class of a
RegisteredExtension. Previously, calling get_extension_class repeatedly
would attempt to load the class mutiple times, but now it's cached.