Rewrite the Manage Extensions page.
The old Manage Extensions page broke with the static media
reorganization and Pipelining, as the enable/disable extension function
was wrapped in a closure and wasn't accessible.

An attempt to fix this spiraled out of control, and now there's a whole
new page. It looks exactly like the old page, but is backed by a set of
JavaScript objects that can call into the extension API to list, enable,
and disable extensions.

We'll be able to build on top of this later on for things like
interfacing with an extension store.

The only difference between the UI in this and the previous version is
that extension dependencies are no longer listed. We haven't even tested
the list since redoing the look of this page, and I'd rather rethink
later on how best to really represent this. As it is now, this shouldn't
really impact anybody.

Some of the CSS for this page was brought over from Review Board as
well, helping to keep the code all in one place.