Enable use of Django admin widgets from extension configuration pages.
The Django admin forms have some very useful widgets that one may want
to pull in. However, it also needed some additional files imported and
variables declared first. It's now possible to use these widgets.

It also makes it possible to define a Media class in a form and include
that media in any siteconfig settings page. This is a standard form
feature we just weren't supporting.

This does require a 'scripts-post' block in a base template to use this
support.