Add a signal to notify when an extension's settings are saved.
A pattern with extensions was to have an extension's SettingsForm
subclass call into the Extension when the form saved, in order to
operate on the new values. However, that only really covers when
SettingsForm saved, and wasn't really the best way to do things.

This change introduces a new signal, settings_saved, which is emitted
whenever an extension's settings have been saved. Extensions can listen
to this, with 'self' as the sender, in order to operate on any new
settings.