Backport the unicode changes for RegistrationForm.
The recent unicode cleanup removed some ancient form encoding logic in
RegistrationForm. This logic wasn't really necessary nowadays, and
mostly just caused trouble, as it expected every field on the form was a
string. That made it difficult to subclass and use things like
dictionaries or lists.

This backports just the part of the code that deals with fetching fields
from the form.