Initialize Lazy Strings before passing to json.dumps
When trying to view the Repository Admin page, the master branch throws a
"Object is not serializable" error because the value getting passed is a
Promise (django.utils.functional.Promise) object rather than a decodable
string. Create a LazyEncoder that will check for this type and call force_text
on it so that it gets initialized properly. Reference:
https://docs.djangoproject.com/en/dev/topics/serialization/#id2

Testing Done:
WebUI - Ensured that the repository admin page works fine.

Reviewed at https://reviews.reviewboard.org/r/5038/