Properly represent JSON data in text fields in the admin UI.
Our JSON support has historically had some issues in the admin UI,
though we broke it a bit further in a recent release. This change
finally makes it work the way we always dreamed it would.

JSON data is now properly shown as-is, not as a Python repr() of the
structure, complete with validation. What's saved becomes valid JSON in
the model.

To make this work, JSONField now uses a custom form field
(JSONFormField) to do the conversion at the appropriate time from the
Python data to the serialized JSON. This will automatically be used in
every part of the admin UI and in every form that uses a JSONField.