Compile and ship localization files for djblets.
This works just like the build process for reviewboard. When doing an sdist, we
compile the .po files to .mo, and include everything in the locale directory
with the distribution.

Testing done:
Built an egg and verified that the build/lib directory contained a locale
directory that had both .po and .mo files. Installed the egg and checked
the files in site-packages.

Reviewed at https://reviews.reviewboard.org/r/5100/