Add support for checkbox columns to the datagrid.
This introduces CheckboxColumn, which can be used by a datagrid to
handle multiple selection of entries. It will display a checkbox for
every row, as well as a checkbox in the header that, when clicked, will
toggle selection of all the checkboxes in each row.

This is based heavily on on Graeme Coupar's work on
https://reviews.reviewboard.org/r/3796/