Support custom "end" prefixes for @blocktag.
@blocktag no longer hard-codes `end` as the prefix for the end tag. A
caller can pass `end_prefix="end_"` or similar to @blocktag, which is
useful when naming a function with underscores.

Decorators are expected to act differently depending on whether they're
called in `@decorator` form or `@decorator()` form. Rather than breaking
every single usage of @blocktag by requiring parens (since we now have
an optional argument), we have to do some checking of arguments at the
end to determine what to return. This is pretty harmless in this case,
but worth pointing out.